# tilt: mac improvements