--- a/Tiltfile
+++ b/Tiltfile
@@ -39,6 +39,7 @@ config.define_bool("manual", False, "Set TRIGGER_MODE_MANUAL by default")
 config.define_bool("m1", False, "Use this flag for M-series Macs (e.g. use an arm64 solana-test-validator due to AVX requirement)")
 
 config.define_string("num", False, "Number of guardian nodes to run")
+config.define_string("maxWorkers", False, "Maximum number of workers for sdk-ci-tests. See https://jestjs.io/docs/cli#--maxworkersnumstring")
 
 # You do not usually need to set this argument - this argument is for debugging only. If you do use a different
 # namespace, note that the "wormhole" namespace is hardcoded in tests and don't forget specifying the argument
@@ -78,6 +79,7 @@ config.define_bool("query_server", False, "Enable cross-chain query server")
 
 cfg = config.parse()
 num_guardians = int(cfg.get("num", "1"))
+max_workers = cfg.get("maxWorkers", "50%")
 namespace = cfg.get("namespace", "wormhole")
 webHost = cfg.get("webHost", "localhost")
 ci = cfg.get("ci", False)
@@ -644,8 +646,10 @@ if ci_tests:
     k8s_yaml_with_ns(
         encode_yaml_stream(
             set_env_in_jobs(
-                set_env_in_jobs(read_yaml_stream("devnet/tests.yaml"), "NUM_GUARDIANS", str(num_guardians)),
-                "BOOTSTRAP_PEERS", str(ccqBootstrapPeers)))
+                set_env_in_jobs(
+                    set_env_in_jobs(read_yaml_stream("devnet/tests.yaml"), "NUM_GUARDIANS", str(num_guardians)),
+                    "BOOTSTRAP_PEERS", str(ccqBootstrapPeers)),
+                    "MAX_WORKERS", max_workers))
     )
 
     # separate resources to parallelize docker builds