--- a/testing/sdk.sh
+++ b/testing/sdk.sh
@@ -8,4 +8,4 @@ done
 while [[ "$(curl -s -o /dev/null -w ''%{http_code}'' spy:6060/metrics)" != "200" ]]; do sleep 5; done
 while [[ "$(curl -s -o /dev/null -w ''%{http_code}'' ibc-relayer:7597/debug/pprof/)" != "200" ]]; do sleep 5; done
 while [[ "$(curl -s -o /dev/null -w ''%{http_code}'' relayer-engine:3000/metrics)" != "200" ]]; do sleep 5; done
-CI=true npm --prefix ../sdk/js run test-ci
+CI=true npm --prefix ../sdk/js run test-ci -- --maxWorkers=$MAX_WORKERS