@evan-gray 2024-12-06
Docker resource usage. It may be worth exposing a parameter for `--maxWorkers`, but I think a better approach would be to remove this deprecated SDK and rely on the new SDK and newer tooling (such as Bun) for the equivalent tests.

<img width="340" alt="image" src="https://github.com/user-attachments/assets/bf5e4bcc-2d7b-49d4-ab01-e0edb6010a97">